Fix: 물놀이 장소 리뷰 조회시 이미지 조회 안되는 오류 해결
